Configies generate_hashed_password
