+++ config/config.exs
@@ -18,3 +18,7 @@ else
 end
 
 config :ex_unit, :capture_log, true
+
+config :authsense,
+  Authsense.Test.User,
+  repo: Authsense.Test.Repo
