+++ lib/authsense/service.ex
@@ -50,13 +50,34 @@ defmodule Authsense.Service do
   end
 
   @doc """
-  See `Authsense.generate_hashed_password/1`.
+  Updates an `Ecto.Changeset` to generate a hashed password.
+  
+  If the changeset has `:password` in it, it will be hashed and stored as
+  `:hashed_password`.  (Fields can be configured in `Authsense`.)
+
+      changeset
+      |> Auth.generate_hashed_password()
+
+  It's typically used in a model's `changeset/2` function.
+
+      defmodule Example.User do
+        use Example.Web, :model
+
+        def changeset(model, params \\ :empty) do
+          model
+          |> cast(params, @required_fields, @optional_fields)
+          |> Auth.generate_hashed_password()
+          |> validate_confirmation(:password, message: "password confirmation doesn't match")
+          |> unique_constraint(:email)
+        end
+      end
+
+  Also see `Authsense.Service.generate_hashed_password/2` for the underlying
+  implementation.
   """
-  def generate_hashed_password(
-    %Ecto.Changeset{} = changeset,
+  def generate_hashed_password(%Ecto.Changeset{} = changeset) do
     %{password_field: passwd, hashed_password_field: hashed_passwd,
-      crypto: crypto})
-  do
+      crypto: crypto} = Authsense.config
     case get_change(changeset, passwd) do
       nil ->
         changeset
